test: Split large tests into smaller compile units
